+++ crates/rpc/rpc-eth-api/Cargo.toml
@@ -29,9 +29,8 @@ reth-execution-types.workspace = true
 reth-rpc-eth-types.workspace = true
 reth-rpc-server-types.workspace = true
 reth-network-api.workspace = true
-reth-trie.workspace = true
 reth-node-api.workspace = true
-
+reth-trie.workspace = true
 # ethereum
 alloy-serde.workspace = true
 alloy-eips.workspace = true
