+++ crates/trie/common/src/lib.rs
@@ -1,5 +1,6 @@
 //! Commonly used types for trie usage.
 
+#![allow(unused_crate_dependencies)]
 #![doc(
     html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
     html_favicon_url = "https://avatars0.githubusercontent.com/u/97369466?s=256",
